FUlly reorganised theory files. All seems functionning. Theory files now all simply have to extend proven.Main, and get every import and tool. Documentation for recent development. Adaptation of existing Set Theory proofs to the new canonical style.
