+++ src/main/scala/utilities/Helpers.scala
@@ -1,3 +1,10 @@
 package utilities
 
+/**
+ * A helper file that provides various syntactic sugars for LISA's FOL and proofs.
+ * Usage:
+ * <pre>
+ * import utilities.Helpers.*
+ * </pre>
+ */
 object Helpers extends TheoriesHelpers {}
